Allow extensions to patch CatnipOptions.
When an extension is starting, it has the option to use
Extension#injectOptions to change the options provided to the catnip
instance. The function takes in a copy of the options object and does
any necessary changes, then returns it. The catnip instance then applies
this option set. The token and extension-override-logging CANNOT be
patched with this.