Convert routes to :params instead of {params}
The reasoning behind this is that, while it's nice to have the routes
follow the same format as the official docs, it's also a bit annoying to
have `{}` in the parameters, as it makes things like using the route
objects directly to make a REST proxy or similar a bit annoying. By
changing to use `:params` instead, the routes become more in-line with
what many common libraries/frameworks expect, and don't require any
pre/post-processing in order to be usable (ex. because a library doesn't
allow for `{}` in routes, or only uses `:params` for dynamic route
params).