Force HTTP/1.1 for default HTTP client.
This is a **temporary** hack to fix #417. We can avoid the specific bug that shows up in that issue by forcing HTTP/1.1 instead of HTTP/2 so that we straight-up never run into that issue. This is a NON-OPTIMAL solution, and as such DOES NOT CLOSE #417, but it WORKS in the short term. Realistically, this will be solved by using OkHTTP or something, if we want to use HTTP/2, but in THEORY this should work for now.